requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.3
django-allauth==65.2.0
django-widget-tweaks==1.5.0
idna==3.10
oauthlib==3.2.2
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
razorpay==1.4.2
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
